backend/requirements.txt
flask
python-dotenv
psycopg2-binary
sqlalchemy-orm
flask_sqlalchemy
flask_migrate
